[dev] Version 2101.2.1-build.39 feat(core): 支持离线提及历史记录并按偏好清理
- MentionRules 新增 allow_offline_history；player/ftb_team 提及启用离线记录
- MentionHistoryRecorder 合并离线目标并新增 maxOfflineTargetsPerMention 限制
- MentionRecord 增加 mentionTypeId，玩家登录时按偏好清理不允许记录

refactor(api): TargetProvider 增加离线目标获取

- PlayerNameTargetProvider/FTBTeamTargetProvider 补充离线 UUID 来源

fix(core): 离线玩家提及解析补全

- MentionResolver 使用 OfflinePlayerList 识别离线玩家名

refactor(client): 历史/提示与候选优化

- MentionHistoryRow 统一 ClickEvent 遍历逻辑
- MentionSuggestions 去重候选并减少在线列表刷新
- ClientMentionContext 缺失 registry 仅警告一次
- MentionPreferencesScreen 修正 Toast 位置计算的 float 兼容

chore(core): 细节优化

- MentionTokens 空匹配提前返回
- NetworkHandler 统一历史开关判定
- MentionSavedData 关闭裁剪时跳过 prune
- FTBChunks 临时路点命令空白分割更稳健
- MentionExecutionSupport 统一 item mention 判定

feat(core): Support offline mention history with preference cleanup

- MentionRules adds allow_offline_history; player/ftb_team enable offline history
- MentionHistoryRecorder merges offline targets and adds maxOfflineTargetsPerMention cap
- MentionRecord adds mentionTypeId so login can prune disallowed records

refactor(api): TargetProvider adds offline target lookup

- PlayerNameTargetProvider/FTBTeamTargetProvider supply offline UUIDs

fix(core): Resolve offline player mentions

- MentionResolver checks OfflinePlayerList for player mentions

refactor(client): History/tooltip and suggestion tweaks

- MentionHistoryRow unifies ClickEvent traversal
- MentionSuggestions de-duplicates candidates and reduces online list refresh
- ClientMentionContext warns once when registry is missing
- MentionPreferencesScreen adjusts toast top positioning with float

chore(core): Small cleanups

- MentionTokens returns early when no matches
- NetworkHandler centralizes history-enabled checks
- MentionSavedData skips prune when disabled
- FTBChunks transient command splits on whitespace
- MentionExecutionSupport centralizes item mention detection

+++ src/main/java/cn/qihuang02/callyou/core/client/screen/MentionPreferencesScreen.java
@@ -292,7 +292,7 @@ public class MentionPreferencesScreen extends ModularUIScreen {
         int left = Math.max(0, (PANEL_WIDTH - toastWidth) / 2);
         toastLayer.layout(style -> style.positionType(YogaPositionType.ABSOLUTE)
                 .left(left)
-                .top((PANEL_HEIGHT - TOAST_HEIGHT) / 2)
+                .top((float) (PANEL_HEIGHT - TOAST_HEIGHT) / 2)
                 .width(toastWidth)
                 .height(TOAST_HEIGHT)
                 .alignItems(YogaAlign.STRETCH)
@@ -716,7 +716,7 @@ public class MentionPreferencesScreen extends ModularUIScreen {
             UIElement toastLayer = new UIElement()
                     .layout(style -> style.positionType(YogaPositionType.ABSOLUTE)
                             .left(0)
-                            .top((PANEL_HEIGHT - TOAST_HEIGHT) / 2)
+                            .top((float) (PANEL_HEIGHT - TOAST_HEIGHT) / 2)
                             .height(TOAST_HEIGHT)
                             .alignItems(YogaAlign.STRETCH)
                             .justifyItems(YogaJustify.CENTER)
